move get_atol_rtol_pcc function to util
